+++ lib/router.js
@@ -1561,7 +1561,7 @@ router.get('/mit/csail/news', lazyloadRouteHandler('./routes/universities/mit/cs
 router.get('/kpmg/insights', lazyloadRouteHandler('./routes/kpmg/insights'));
 
 // Saraba1st
-router.get('/saraba1st/thread/:tid', lazyloadRouteHandler('./routes/saraba1st/thread'));
+// router.get('/saraba1st/thread/:tid', lazyloadRouteHandler('./routes/saraba1st/thread'));
 
 // gradcafe
 router.get('/gradcafe/result/:type', lazyloadRouteHandler('./routes/gradcafe/result'));
