+++ lib/v2/saraba1st/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/thread/:tid': ['zengxs'],
+};
