+++ lib/v2/saraba1st/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/thread/:tid', require('./thread'));
+};
