+++ lib/v2/saraba1st/thread.js
@@ -1,4 +1,3 @@
-const url = require('url');
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
 const queryString = require('query-string');
@@ -49,7 +48,7 @@ module.exports = async (ctx) => {
         contentHtml.find('div.aimg_tip').remove();
         return {
             title: `${title} #${floor}`,
-            link: url.resolve('https://bbs.saraba1st.com/2b/', floorUrl),
+            link: new URL(floorUrl, 'https://bbs.saraba1st.com/2b/').href,
             description: contentHtml.html(),
             author: each.find('.authi .xw1').text(),
             pubDate: timezone(parseDate(each.find('.authi em').text()), +8),
